requirements.txt
Flask==1.1.2
ffmpeg-python==0.2.0
gunicorn==20.0.4
flask-cors==3.0.9